Organize file structure and refactor components
This commit refactors the Page component by accepting a children prop. Because all other components related to different pages of the application share components such as the header and footer, the Page component is called with a child component of a specific page component as a child in Router to render the correct page.