Load and display products of a given category and create a product page for individual products
This commit refactors the AllProductsPage component to map ProductCard components (that display individual product cards) based on the data retrieved from an API fetch. Because of the dynamic routes determined in the Router component, the AllProductsPage component also renders the ProductPage component when a single product card is clicked on.

This commit also removes the Page component to make routing logic easier to maintain. Styling from the Page component was applied to components that were children of the Page component as well as index.html.