Refactor App component to handle routing logic and implement feature to track number of items in the bag
This commit refactors the App component by moving the original logic from the Router component to the App component itself. This reduces the number of components and simplifies the flow of logic while maintaining separation of concerns. Test files that rendered the App component were updated to render the HomePage component to address the error of rendering a <Router> inside another <Router>.

This commit also implements a function that keeps track of the products added to the shopping bag. Whenever a product is added to the bag, the header will display the total number of items in the bag. A future implementation will include updating the shopping bag component to display all products inside of the bag.