Fix incorrect number of items added to the bag displayed in the header and inability to add more than two quantities of a single product
This commit displays the same number of items currently in the bag when the header is re-rendered across different components. A 'quantity' key was added to a new instance of the fetched product data to keep track of individual and total item quantities. A 'key' prop was also included to the AllProductsPage component when rendering multiple path routes using the same element; this helped differentiate each element to React and allow the user to add more than two quantities of a single product to the bag.