Create a sidebar menu for smaller screens and allow the user to open or close it
This commit adds a sidebar feature to the application that contains navigation links to other pages. These links are the same ones displayed on the header for larger screens, but for smaller screens, the links are replaced with a mobile menu icon. Upon clicking the mobile menu icon, the sidebar opens, which allows the user to find the same navigation links that were originally displayed on the header.